requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14
psycopg[binary]>=3.1
python-dotenv>=1.0
djangorestframework-simplejwt>=5.4.0,<6.0
django-filter>=23.0
pytest>=7.0
pytest-django>=4.5
drf-spectacular>=0.29.0,<0.30.0
drf-spectacular-sidecar>=2025.12.1,<2026.0.0
